+++ src/main/java/io/CarNameInputHandler.java
@@ -1,4 +1,4 @@
-package iohandler;
+package io;
 
 import java.util.Arrays;
 import java.util.List;
