+++ src/test/java/io/CarNameInputHandlerTest.java
@@ -1,4 +1,4 @@
-package iohandler;
+package io;
 
 import org.junit.jupiter.api.Test;
 
@@ -6,7 +6,6 @@ import java.util.List;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
-import static org.junit.jupiter.api.Assertions.*;
 
 class CarNameInputHandlerTest {
